confirmation message done

+++ src/main/java/com/engly/engly_server/service/impl/NotificationServiceImpl.java
@@ -1,7 +1,10 @@
 package com.engly.engly_server.service.impl;
 
 
+import com.engly.engly_server.exception.NotFoundException;
+import com.engly.engly_server.models.entity.Users;
 import com.engly.engly_server.models.entity.VerifyToken;
+import com.engly.engly_server.repo.UserRepo;
 import com.engly.engly_server.repo.VerifyTokenRepo;
 import com.engly.engly_server.service.EmailService;
 import com.engly.engly_server.service.NotificationService;
@@ -20,19 +23,25 @@ public class NotificationServiceImpl implements NotificationService {
 
     private final EmailService emailService;
     private final EmailMessageGenerator messageGenerator;
+    private final UserRepo userRepo;
 
     private static final Logger log = LoggerFactory.getLogger(NotificationServiceImpl.class);
 
 
-    public NotificationServiceImpl(VerifyTokenRepo tokenRepo, EmailService emailService, EmailMessageGenerator messageGenerator) {
+    public NotificationServiceImpl(VerifyTokenRepo tokenRepo, EmailService emailService, EmailMessageGenerator messageGenerator, UserRepo userRepo) {
         this.tokenRepo = tokenRepo;
         this.emailService = emailService;
         this.messageGenerator = messageGenerator;
+        this.userRepo = userRepo;
     }
 
 
     public void sendNotifyMessage(String email) {
         try {
+            if (!userRepo.existsByEmail(email)) {
+                throw new NotFoundException("User not found exception email %s".formatted(email));
+            }
+
             String token = RandomStringUtils.random(32, true, true);
             tokenRepo.save(new VerifyToken(token, email));
 
@@ -52,12 +61,17 @@ public class NotificationServiceImpl implements NotificationService {
     public void checkToken(String token, String email) {
         Optional<VerifyToken> verifyToken = tokenRepo.findByTokenAndEmail(token, email);
 
-        if (verifyToken.isPresent()) {
+        Optional<Users> user = userRepo.findByEmail(email);
+
+        if (verifyToken.isPresent() && user.isPresent()) {
+            Users users = user.get();
+            users.setEmailVerified(true);
+
             tokenRepo.delete(verifyToken.get());
             log.info("[NotificationServiceImpl:checkToken]Token:{} for email:{} was checked and deleted", token, email);
         } else {
             log.info("[NotificationServiceImpl:checkToken]Token:{} for email:{} was not verified", token, email);
-            throw new RuntimeException("Token not found exception email %s".formatted(email));
+            throw new NotFoundException("Token not found exception email %s".formatted(email));
         }
     }
 }
